+++ common/platform.h
@@ -30,15 +30,15 @@ extern Model model_vm;
 void read_from_nvm(void* vm_buffer, uint32_t nvm_offset, size_t n);
 void write_to_nvm(const void* vm_buffer, uint32_t nvm_offset, size_t n, uint16_t timer_delay = 0);
 // DMA controller on MSP432 can handle at most 1024 words at a time
-void write_to_nvm_segmented(const uint8_t* vm_buffer, uint32_t nvm_offset, uint16_t total_len, uint16_t segment_size = 1024);
+void write_to_nvm_segmented(const uint8_t* vm_buffer, uint32_t nvm_offset, uint32_t total_len, uint16_t segment_size = 1024);
 void my_erase(void);
 void copy_samples_data(void);
 void my_memcpy(void* dest, const void* src, size_t n);
-void my_memcpy_to_param(ParameterInfo *param, uint16_t offset_in_word, const void *src, size_t n, uint16_t timer_delay);
-void my_memcpy_from_intermediate_values(void *dest, const ParameterInfo *param, uint16_t offset_in_word, size_t n);
+void my_memcpy_to_param(ParameterInfo *param, uint32_t offset_in_word, const void *src, size_t n, uint16_t timer_delay);
+void my_memcpy_from_intermediate_values(void *dest, const ParameterInfo *param, uint32_t offset_in_word, size_t n);
 // offset_in_bytes may go beyond 64K after being multiplied with sizeof(T)
 void my_memcpy_from_parameters(void *dest, const ParameterInfo *param, uint32_t offset_in_bytes, size_t n);
-void read_from_samples(void *dest, uint16_t offset_in_word, size_t n);
+void read_from_samples(void *dest, uint32_t offset_in_word, size_t n);
 ParameterInfo* get_intermediate_parameter_info(uint8_t i);
 void commit_intermediate_parameter_info(uint8_t i);
 Model* get_model(void);
