+++ common/pooling.cpp
@@ -105,7 +105,7 @@ static uint16_t maxpool_patch(MaxPoolParams *maxpool_params) {
             if (input_h >= maxpool_params->H || input_w >= maxpool_params->W) {
                 continue;
             }
-            uint16_t val_offset = input_h * offset_h + input_w * offset_w + maxpool_params->start_channel;
+            uint32_t val_offset = input_h * offset_h + input_w * offset_w + maxpool_params->start_channel;
             my_memcpy_from_param(maxpool_params->model, input_buffer, maxpool_params->data, val_offset, maxpool_params->n_channels * sizeof(int16_t));
             output_channel_offset = 0;
             for (uint16_t input_channel_offset = 0; input_channel_offset < maxpool_params->n_channels; input_channel_offset++) {
@@ -144,7 +144,7 @@ static uint16_t maxpool_patch(MaxPoolParams *maxpool_params) {
 }
 
 #if STATEFUL
-static inline void offset_vector(int16_t* const buffer, int16_t offset, uint8_t len, const uint16_t output_offset, const uint16_t next_output_turning_point) {
+static inline void offset_vector(int16_t* const buffer, int16_t offset, uint8_t len, const uint32_t output_offset, const uint16_t next_output_turning_point) {
     int16_t cur_offset = offset;
     for (uint8_t idx = BATCH_SIZE - 1; idx < len; idx += BATCH_SIZE) {
         if (output_offset + idx == next_output_turning_point + BATCH_SIZE - 1) {
@@ -155,7 +155,7 @@ static inline void offset_vector(int16_t* const buffer, int16_t offset, uint8_t 
 }
 #endif
 #if JAPARI
-static inline void offset_vector(int16_t* const buffer, int16_t offset, uint8_t len, const uint16_t output_offset, const uint16_t next_output_turning_point) {
+static inline void offset_vector(int16_t* const buffer, int16_t offset, uint8_t len, const uint32_t output_offset, const uint16_t next_output_turning_point) {
     int16_t cur_footprint = (offset == 0x4000 ? 1 : -1);
     uint8_t reverted = 0;
     for (uint8_t idx = BATCH_SIZE; idx < len; idx += BATCH_SIZE + 1) {
@@ -182,7 +182,7 @@ void handle_maxpool(Model *model, const ParameterInfo *input[], ParameterInfo *o
     const uint16_t CHANNEL = data->dims[1], OUTPUT_CHANNEL = output->dims[1];
 
     uint16_t output_h = 0, output_w = 0, c = 0;
-    uint16_t output_offset = 0;
+    uint32_t output_offset = 0;
 
 #if INTERMITTENT
     start_cpu_counter(offsetof(Counters, progress_seeking));
